+++ CHANGELOG.md
@@ -17,7 +17,8 @@ changes.
 
 ### Fixed
 
--
+- Fix querying metadata-validation service from haskell backend [Issue 1612](https://github.com/IntersectMBO/govtool/issues/1612)
+- Fix thread killed by timeout manager [Issue 1417](https://github.com/IntersectMBO/govtool/issues/1417)
 
 ### Changed